Added logging and locking of sleep if a server is still running

+++ allsparkmaster_bot.py
@@ -38,8 +38,7 @@ updater = Updater(token=strBotToken)
 
 dispatcher = updater.dispatcher
 
-logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-                     level=logging.INFO)
+logging.basicConfig(filename='cheetor.log', format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
 
 ssh = paramiko.SSHClient()
 ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
@@ -58,14 +57,64 @@ def setThreadUnlocked():
     global bThreadLocked
     bThreadLocked = False
 
+# Server running functions
+bArkserver = False
+bCssserver = False
+bMcserver = False
+
+def setArkserverStarted():
+    logging.debug(setArkserverStarted.__name__)
+    global bArkserver
+    bArkserver = True
+
+def setArkserverStopped():
+    logging.debug(setArkserverStopped.__name__)
+    global bArkserver
+    bArkserver = False
+
+def getArkserverStatus():
+    logging.debug(getArkserverStatus.__name__)
+    return bArkserver
+
+def setCssserverStarted():
+    logging.debug(setCssserverStarted.__name__)
+    global bCssserver
+    bCssserver = True
+
+def setCssserverStopped():
+    logging.debug(setCssserverStopped.__name__)
+    global bCssserver
+    bCssserver = False
+
+def getCssserverStatus():
+    logging.debug(getCssserverStatus.__name__)
+    return bCssserver
+
+def setMcserverStarted():
+    logging.debug(setMcserverStarted.__name__)
+    global bMcserver
+    bMcserver = True
+
+def setMcserverStopped():
+    logging.debug(setMcserverStopped.__name__)
+    global bMcserver
+    bMcserver = False
+
+def getMcserverStatus():
+    logging.debug(getMcserverStatus.__name__)
+    return bMcserver
+
 # Functions for handlers
 def help(update, context):
+    logging.debug(help.__name__)
     context.bot.send_message(chat_id=update.effective_chat.id, text=strHelp)
 
 def echo(update, context):
+    logging.debug(echo.__name__)
     context.bot.send_message(chat_id=update.effective_chat.id, text=strHelp)
 
 def wakeUpAllspark(update, context):
+    logging.info(wakeUpAllspark.__name__)
     if False == getThreadLocked():
         setThreadLocked()
         send_magic_packet(strMac)
@@ -73,54 +122,71 @@ def wakeUpAllspark(update, context):
         time.sleep(30)
         retValue = connectSsh(strIp, strCommodus, strCommodusPw, update, context)
         if True == retValue:
-            context.bot.send_message(chat_id=update.effective_chat.id, text="The Allspark is awake")
+            str = "The Allspark is awake"
+            logging.info(str)
+            context.bot.send_message(chat_id=update.effective_chat.id, text=str)
         else:
-            context.bot.send_message(chat_id=update.effective_chat.id, text="ERROR: The Allspark is still sleeping")
+            str = "ERROR: The Allspark is still sleeping"
+            logging.info(str)
+            context.bot.send_message(chat_id=update.effective_chat.id, text=str)
         ssh.close()
         setThreadUnlocked()
     else:
-        context.bot.send_message(chat_id=update.effective_chat.id, text="Allspark is busy, wait for last command to finish")
+        str = "Allspark is busy, wait for last command to finish"
+        logging.info(str)
+        context.bot.send_message(chat_id=update.effective_chat.id, text=str)
 
 def sleepAllspark(update, context):
+    logging.info(sleepAllspark.__name__)
     if False == getThreadLocked():
         setThreadLocked()
-        context.bot.send_message(chat_id=update.effective_chat.id, text="The Allspark will sleep!")
-        retValue = connectSsh(strIp, strCommodus, strCommodusPw, update, context)
-        if True == retValue:
-            stdin, stdout, stderr = ssh.exec_command('sudo shutdown -P now', get_pty=True)
-            stdin.write(strCommodusPw)
-            stdin.write("\n")
-            stdin.flush()
-        ssh.close()
+        if (False == getArkserverStatus()) and (False == getCssserverStatus()) and (False == getMcserverStatus()):
+            context.bot.send_message(chat_id=update.effective_chat.id, text="The Allspark will sleep!")
+            retValue = connectSsh(strIp, strCommodus, strCommodusPw, update, context)
+            if True == retValue:
+                stdin, stdout, stderr = ssh.exec_command('sudo shutdown -P now', get_pty=True)
+                time.sleep(1)
+                stdin.write(strCommodusPw)
+                stdin.write("\n")
+                time.sleep(1)
+                stdin.flush()
+            ssh.close()
+        else:
+            context.bot.send_message(chat_id=update.effective_chat.id, text="ERROR: Check running servers!")
         setThreadUnlocked()
     else:
         context.bot.send_message(chat_id=update.effective_chat.id, text="Allspark is busy, wait for last command to finish")
 
 def startArkserver(update, context):
+    logging.info(startArkserver.__name__)
     if False == getThreadLocked():
         setThreadLocked()
         context.bot.send_message(chat_id=update.effective_chat.id, text="Starting Arkserver")
         retValue = connectSsh(strIp, strArkserver, strArkserverPw, update, context)
         if True == retValue:
             sendCommand('./arkserver start', update, context)
+            setArkserverStarted()
         ssh.close()
         setThreadUnlocked()
     else:
         context.bot.send_message(chat_id=update.effective_chat.id, text="Allspark is busy, wait for last command to finish")
 
 def stopArkserver(update, context):
+    logging.info(stopArkserver.__name__)
     if False == getThreadLocked():
         setThreadLocked()
         context.bot.send_message(chat_id=update.effective_chat.id, text="Stopping Arkserver")
         retValue = connectSsh(strIp, strArkserver, strArkserverPw, update, context)
         if True == retValue:
             sendCommand('./arkserver stop', update, context)
+            setArkserverStopped()
         ssh.close()
         setThreadUnlocked()
     else:
         context.bot.send_message(chat_id=update.effective_chat.id, text="Allspark is busy, wait for last command to finish")
 
 def statusArkserver(update, context):
+    logging.info(statusArkserver.__name__)
     if False == getThreadLocked():
         setThreadLocked()
         context.bot.send_message(chat_id=update.effective_chat.id, text="Getting status of Arkserver")
@@ -138,48 +204,56 @@ def statusArkserver(update, context):
         context.bot.send_message(chat_id=update.effective_chat.id, text="Allspark is busy, wait for last command to finish")
 
 def startCssserver(update, context):
+    logging.info(startCssserver.__name__)
     if False == getThreadLocked():
         setThreadLocked()
         context.bot.send_message(chat_id=update.effective_chat.id, text="Starting Cssserver!")
         retValue = connectSsh(strIp, strCssserver, strCssserverPw, update, context)
         if True == retValue:
             sendCommand('./cssserver start', update, context)
+            setCssserverStarted()
         ssh.close()
         setThreadUnlocked()
     else:
         context.bot.send_message(chat_id=update.effective_chat.id, text="Allspark is busy, wait for last command to finish")
 
 def stopCssserver(update, context):
+    logging.info(stopCssserver.__name__)
     if False == getThreadLocked():
         setThreadLocked()
         context.bot.send_message(chat_id=update.effective_chat.id, text="Stopping Cssserver!")
         retValue = connectSsh(strIp, strCssserver, strCssserverPw, update, context)
         if True == retValue:    
             sendCommand('./cssserver stop', update, context)
+            setCssserverStopped()
         ssh.close()
         setThreadUnlocked()
     else:
         context.bot.send_message(chat_id=update.effective_chat.id, text="Allspark is busy, wait for last command to finish")
 
 def startMcserver(update, context):
+    logging.info(startMcserver.__name__)
     if False == getThreadLocked():
         setThreadLocked()
         context.bot.send_message(chat_id=update.effective_chat.id, text="Starting Minecraft server")
         retValue = connectSsh(strIp, strMcserver, strMcserverPw, update, context)
         if True == retValue:
             sendCommand('./mcserver start', update, context)
+            setMcserverStarted()
         ssh.close()
         setThreadUnlocked()
     else:
         context.bot.send_message(chat_id=update.effective_chat.id, text="Allspark is busy, wait for last command to finish")
 
 def stopMcserver(update, context):
+    logging.info(stopMcserver.__name__)
     if False == getThreadLocked():
         setThreadLocked()
         context.bot.send_message(chat_id=update.effective_chat.id, text="Stopping Minecraft server")
         retValue = connectSsh(strIp, strMcserver, strMcserverPw, update, context)
         if True == retValue:
             sendCommand('./mcserver stop', update, context)
+            setMcserverStopped()
         ssh.close()
         setThreadUnlocked()
     else:
@@ -187,6 +261,7 @@ def stopMcserver(update, context):
 
 
 def inline(update, context):
+    logging.info(inline.__name__)
     query = update.inline_query.query
     if not query:
         return
